pyproject.toml
[tool.poetry]
name = "bibloteca-digital-bogota"
version = "0.1.0"
description = ""
authors = ["Aguayo <contacto@aguayo.co>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4"
bleach = "^3.1"
boto3 = "^1.9"
django-ckeditor = "^6.7.1"
django-constance = "^3.1.0"
django-db-logger = "^0.1.13"
django-debug-toolbar = "^4.4.2"
django-extensions = "^3.2.3"
django-formtools = "^2.1"
django-google-tag-manager = "^0.0.5"
django-heroku = "^0.3.1"
django-import-export = "4.1.1"
django-q = "^1.3.9"
django-simple-history = "3.3.0"
django-simple-robots = "^1.0"
django-storages = "^1.14.3"
django-webpack-loader = "^0.6.0"
django-xff = "^1.3"
docutils = "^0.15.2"
gunicorn = "^20.0.4"
ipaddress = "^1.0"
patool = "^1.12" # Used by pyunpack
pillow = "^9.0"
psycopg2 = "^2.7"
pymarc = "^3.1"
pyunpack = "^0.2.2"
requests = "^2.22"
roman-numerals = "^0.1.0"
scout-apm = {version = "^2.5", platform = "linux || darwin"}
sickle = "^0.6.4"
tomlkit = "^0.5.3"
wagtail = "6.1.2"
wagtail-instance-selector = "2.1.2"
wagtail-modelchooser = "2.14.1"
certifi = "^2020.4.5"
django-select2 = "^7.4.2"
django-gm2m = "^1.2"
django-advanced-filters = "2.0.0"
django-redis = "^4.12.1"
wagtail-cache = "^1.0.1"
django-cache-machine = "^1.1.0"
elasticsearch = "7.13"
Pympler = "^0.9"
memcache_toolbar = "^0.5.6"
django-bmemcached = "^0.3.0"
django-debug-toolbar-template-profiler = "^2.0.2"
django-brotli = "^0.2.0"
elastic-apm = "5.8.1"
wagtailmenus = "^4.0"
six = "^1.16.0"
django-utils-six = "^2.0"
djangorestframework = "^3.14.0"
django-summernote = "^0.8.20.0"
update = "^0.0.1"
wagtail-modeladmin = "^2.0.0"

[tool.poetry.dev-dependencies]
notebook = "^6.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
